+++ src/content-loader/__tests__/validator.js
@@ -5,14 +5,14 @@ test("should pass if no config provided", () => {
   validator()
 })
 
-test("should exit if provide unknow option", (t) => {
+test("should throw if provide unknow option", (t) => {
   t.throws(
     () => validator({ foo: "bar" }),
     (error) => error.message.includes(`Unknow option 'foo'`)
   )
 })
 
-test("should exit if provide wrong type", (t) => {
+test("should throw if provide wrong type", (t) => {
   t.throws(
     () => validator({ context: false }),
     (error) => error.message.includes(`Wrong type for 'context'`)